Some minor changes

+++ src/ann_regression/ann_matlab/+ann_engine/AnnEnginePythonAnn.m
@@ -109,7 +109,7 @@ classdef AnnEnginePythonAnn < ann_engine.AnnEngineAbstract
             assert(data_out.status==true, 'prediction Python error')
             
             % response data
-            out = data_out.out;
+            out = double(data_out.out);
         end
     end
 end
